#129 formData에 File 추가
- 기존 data 객체에서 FormData 형태로 변경
