+++ client/src/components/imageUpload/ImageUpload.tsx
@@ -1,7 +1,8 @@
 import { ChangeEvent } from 'react';
-import tw from 'twin.macro';
 import { styled } from 'styled-components';
+import tw from 'twin.macro';
 
+import { changeImageFileName, createImageDataUrl } from '../../utils/image';
 import { ImgLabel } from '../label/Label';
 import ProfileImg from '../../img/profile_img2.png';
 import Button from '../buttons/Button';
@@ -22,49 +23,9 @@ interface IProps {
 
 const ImageUpload = ({ selectImg, nickname, handleSelectImage, handleFileInfo }: IProps) => {
   const handleImgControl = (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files;
-    const maxSize = 2 * 1024 * 1024;
-
-    if (file && file[0]?.size > maxSize) {
-      alert('이미지 파일은 2MB 이하로만 첨부 가능합니다. :(');
-    }
-
-    /**
-     * make dataURL
-     */
-    if (file && file[0]?.size < maxSize) {
-      const fileReader = new FileReader();
-      fileReader.readAsDataURL(file[0]);
-
-      fileReader.onload = () => {
-        const imgURL = fileReader.result;
-        if (imgURL) {
-          handleSelectImage(imgURL.toString());
-        }
-      };
-    }
-
-    /**
-     * file name change
-     *  - 파일 이름 형식: [nickname]_년월일시분초.[jpg]
-     */
-    if (file) {
-      // 기존 파일
-      const prevFile = file[0];
-
-      // 파일 이름 형식의 `년월일시분초` 생성
-      const date = new Date();
-      const fileName = `${nickname}_${date.getFullYear()}${
-        date.getMonth() + 1
-      }${date.getDate()}${date.getHours()}${date.getMinutes()}${date.getSeconds()}`.split('\n')[0];
-
-      // file 확장자
-      const fileExtension = prevFile.type.split('/')[1];
-
-      // file 이름 변경
-      const newFile = new File([prevFile], `${fileName}.${fileExtension}`, { type: prevFile.type });
-      handleFileInfo(newFile);
-    }
+    const file = createImageDataUrl(e, handleSelectImage);
+    const defaultNickname = nickname ?? 'template';
+    if (file) changeImageFileName(file[0], defaultNickname, handleFileInfo);
   };
 
   const handleDeletePreviewImg = (e: MouseEvent) => {
